spaceiz_chat/requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
backports.zoneinfo==0.2.1
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.1
Django==4.0.6
djangorestframework==3.13.1
gunicorn==20.0.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
supervisor==4.2.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
zope.interface==5.4.0
